asserts: separate throw/returned transaction error assertions
